Reuse ManualWitness
