+++ tests/neo.UnitTests/SmartContract/Native/Tokens/UT_NeoToken.cs
@@ -6,6 +6,7 @@ using Neo.IO;
 using Neo.IO.Caching;
 using Neo.Ledger;
 using Neo.Network.P2P.Payloads;
+using Neo.Oracle;
 using Neo.Persistence;
 using Neo.SmartContract;
 using Neo.SmartContract.Native;
@@ -580,7 +581,7 @@ namespace Neo.UnitTests.SmartContract.Native.Tokens
         internal static (bool State, bool Result) Check_Vote(StoreView snapshot, byte[] account, byte[][] pubkeys, bool signAccount)
         {
             var engine = new ApplicationEngine(TriggerType.Application,
-                new Nep5NativeContractExtensions.ManualWitness(signAccount ? new UInt160(account) : UInt160.Zero), snapshot, 0, true);
+                new ManualWitness(signAccount ? new UInt160(account) : UInt160.Zero), snapshot, 0, true);
 
             engine.LoadScript(NativeContract.NEO.Script);
